// src/protocol/indigo/models/SystemParamsV1.ts
export interface CurrencySymbol {
    unCurrencySymbol: string;
}
export type AssetClass = [CurrencySymbol, TokenName];
// ...
export interface TokenName {
    unTokenName: string;
}